Extract form state management into a custom hook
